# src/routes/unregister_request.rs
use std::sync::Arc;

use axum::extract::{Path, State};
use smelter_core::{LateEventPolicy, Pipeline, Timestamp};
use smelter_render::{RegistryType, error::ErrorStack};
use tracing::error;

use crate::{error::ApiError, state::ApiState};

use smelter_api::{InputId, OkResponse, OutputId, RendererId, UnregisterRequest};

use super::Json;

#[utoipa::path(
    post,
    path = "/api/input/{input_id}/unregister",
    operation_id = "unregister_input",
    params(("input_id" = str, Path, description = "Input ID.")),
    responses(
        (status = 200, description = "Input unregistered successfully.", body = OkResponse),
        (status = 400, description = "Bad request.", body = ApiError),
        (status = 404, description = "Input not found.", body = ApiError),
        (status = 500, description = "Internal server error.", body = ApiError),
    ),
    tags = ["unregister_request"],
)]
pub async fn handle_input(
    State(api): State<Arc<ApiState>>,
    Path(input_id): Path<InputId>,
    Json(request): Json<UnregisterRequest>,
) -> Result<Json<OkResponse>, ApiError> {
    match request.schedule_time_ms {
        Some(schedule_time_ms) => {
            let schedule_time = Timestamp::from_secs_f64(schedule_time_ms / 1000.0);
            Pipeline::schedule_event(
                &api.pipeline()?,
                schedule_time,
                LateEventPolicy::Default,
                move |pipeline| {
                    if let Err(err) = pipeline.unregister_input(&input_id.into()) {
                        error!(
                            "Error while running scheduled input unregister for pts {}ms: {}",
                            schedule_time.as_millis(),
                            ErrorStack::new(&err).into_string()
                        )
                    }
                },
            );
        }
        None => {
            api.pipeline()?
                .lock()
                .unwrap()
                .unregister_input(&input_id.into())?;
        }
    }
    Ok(Json(OkResponse {}))
}

#[utoipa::path(
    post,
    path = "/api/output/{output_id}/unregister",
    operation_id = "unregister_output",
    params(("output_id" = str, Path, description = "Output ID.")),
    responses(
        (status = 200, description = "Output unregistered successfully.", body = OkResponse),
        (status = 400, description = "Bad request.", body = ApiError),
        (status = 404, description = "Output not found.", body = ApiError),
        (status = 500, description = "Internal server error.", body = ApiError),
    ),
    tags = ["unregister_request"],
)]
pub async fn handle_output(
    State(api): State<Arc<ApiState>>,
    Path(output_id): Path<OutputId>,
    Json(request): Json<UnregisterRequest>,
) -> Result<Json<OkResponse>, ApiError> {
    match request.schedule_time_ms {
        Some(schedule_time_ms) => {
            let schedule_time = Timestamp::from_secs_f64(schedule_time_ms / 1000.0);
            Pipeline::schedule_event(
                &api.pipeline()?,
                schedule_time,
                LateEventPolicy::Default,
                move |pipeline| {
                    if let Err(err) = pipeline.unregister_output(&output_id.into()) {
                        error!(
                            "Error while running scheduled output unregister for pts {}ms: {}",
                            schedule_time.as_millis(),
                            ErrorStack::new(&err).into_string()
                        )
                    }
                },
            );
        }
        None => {
            api.pipeline()?
                .lock()
                .unwrap()
                .unregister_output(&output_id.into())?;
        }
    }
    Ok(Json(OkResponse {}))
}

#[utoipa::path(
    post,
    path = "/api/shader/{shader_id}/unregister",
    operation_id = "unregister_shader",
    params(("shader_id" = str, Path, description = "Shader ID.")),
    responses(
        (status = 200, description = "Shader unregistered successfully.", body = OkResponse),
        (status = 400, description = "Bad request.", body = ApiError),
        (status = 404, description = "Shader not found.", body = ApiError),
        (status = 500, description = "Internal server error.", body = ApiError),
    ),
    tags = ["unregister_request"],
)]
pub async fn handle_shader(
    State(api): State<Arc<ApiState>>,
    Path(shader_id): Path<RendererId>,
    Json(request): Json<UnregisterRequest>,
) -> Result<Json<OkResponse>, ApiError> {
    match request.schedule_time_ms {
        Some(schedule_time_ms) => {
            let schedule_time = Timestamp::from_secs_f64(schedule_time_ms / 1000.0);
            Pipeline::schedule_event(
                &api.pipeline()?,
                schedule_time,
                LateEventPolicy::Default,
                move |pipeline| {
                    if let Err(err) =
                        pipeline.unregister_renderer(&shader_id.into(), RegistryType::Shader)
                    {
                        error!(
                            "Error while running scheduled shader unregister for pts {}ms: {}",
                            schedule_time.as_millis(),
                            ErrorStack::new(&err).into_string()
                        )
                    }
                },
            );
        }
        None => {
            api.pipeline()?
                .lock()
                .unwrap()
                .unregister_renderer(&shader_id.into(), RegistryType::Shader)?;
        }
    }
    Ok(Json(OkResponse {}))
}

#[utoipa::path(
    post,
    path = "/api/web-renderer/{instance_id}/unregister",
    operation_id = "unregister_web_renderer",
    params(("instance_id" = str, Path, description = "Web renderer ID.")),
    responses(
        (status = 200, description = "Web renderer unregistered successfully.", body = OkResponse),
        (status = 400, description = "Bad request.", body = ApiError),
        (status = 404, description = "Web renderer not found.", body = ApiError),
        (status = 500, description = "Internal server error.", body = ApiError),
    ),
    tags = ["unregister_request"],
)]
pub async fn handle_web_renderer(
    State(api): State<Arc<ApiState>>,
    Path(instance_id): Path<RendererId>,
    Json(request): Json<UnregisterRequest>,
) -> Result<Json<OkResponse>, ApiError> {
    match request.schedule_time_ms {
        Some(schedule_time_ms) => {
            let schedule_time = Timestamp::from_secs_f64(schedule_time_ms / 1000.0);
            Pipeline::schedule_event(
                &api.pipeline()?,
                schedule_time,
                LateEventPolicy::Default,
                move |pipeline| {
                    if let Err(err) =
                        pipeline.unregister_renderer(&instance_id.into(), RegistryType::WebRenderer)
                    {
                        error!(
                            "Error while running scheduled web renderer unregister for pts {}ms: {}",
                            schedule_time.as_millis(),
                            ErrorStack::new(&err).into_string()
                        )
                    }
                },
            );
        }
        None => {
            api.pipeline()?
                .lock()
                .unwrap()
                .unregister_renderer(&instance_id.into(), RegistryType::WebRenderer)?;
        }
    }
    Ok(Json(OkResponse {}))
}

#[utoipa::path(
    post,
    path = "/api/image/{image_id}/unregister",
    operation_id = "unregister_image",
    params(("image_id" = str, Path, description = "Image ID.")),
    responses(
        (status = 200, description = "Image unregistered successfully.", body = OkResponse),
        (status = 400, description = "Bad request.", body = ApiError),
        (status = 404, description = "Image not found.", body = ApiError),
        (status = 500, description = "Internal server error.", body = ApiError),
    ),
    tags = ["unregister_request"],
)]
pub async fn handle_image(
    State(api): State<Arc<ApiState>>,
    Path(image_id): Path<RendererId>,
    Json(request): Json<UnregisterRequest>,
) -> Result<Json<OkResponse>, ApiError> {
    match request.schedule_time_ms {
        Some(schedule_time_ms) => {
            let schedule_time = Timestamp::from_secs_f64(schedule_time_ms / 1000.0);
            Pipeline::schedule_event(
                &api.pipeline()?,
                schedule_time,
                LateEventPolicy::Default,
                move |pipeline| {
                    if let Err(err) =
                        pipeline.unregister_renderer(&image_id.into(), RegistryType::Image)
                    {
                        error!(
                            "Error while running scheduled image unregister for pts {}ms: {}",
                            schedule_time.as_millis(),
                            ErrorStack::new(&err).into_string()
                        )
                    }
                },
            );
        }
        None => {
            api.pipeline()?
                .lock()
                .unwrap()
                .unregister_renderer(&image_id.into(), RegistryType::Image)?;
        }
    }
    Ok(Json(OkResponse {}))
}