# tools/schemas/openapi_specification.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Smelter",
    "description": "Real-time video compositing software",
    "license": {
      "name": "MIT"
    },
    "version": "0.6.0"
  },
  "paths": {
...
    "/api/input/{input_id}/unregister": {
      "post": {
        "tags": [
          "unregister_request"
        ],
        "operationId": "unregister_input",
        "parameters": [
          {
            "name": "input_id",
            "in": "path",
            "description": "Input ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnregisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Input unregistered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Input not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/output/{output_id}/unregister": {
      "post": {
        "tags": [
          "unregister_request"
        ],
        "operationId": "unregister_output",
        "parameters": [
          {
            "name": "output_id",
            "in": "path",
            "description": "Output ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnregisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Output unregistered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Output not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/shader/{shader_id}/unregister": {
      "post": {
        "tags": [
          "unregister_request"
        ],
        "operationId": "unregister_shader",
        "parameters": [
          {
            "name": "shader_id",
            "in": "path",
            "description": "Shader ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnregisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Shader unregistered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Shader not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/web-renderer/{instance_id}/unregister": {
      "post": {
        "tags": [
          "unregister_request"
        ],
        "operationId": "unregister_web_renderer",
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "description": "Web renderer ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnregisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Web renderer unregistered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Web renderer not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/image/{image_id}/unregister": {
      "post": {
        "tags": [
          "unregister_request"
        ],
        "operationId": "unregister_image",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "description": "Image ID.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnregisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Image unregistered successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Image not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "required": [
          "error_code",
          "message",
          "stack",
          "http_status_code"
        ],
        "properties": {
          "error_code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "stack": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "http_status_code": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "OkResponse": {
        "type": "object",
        "description": "Response for requests that do not return any data."
      },
...
      "UnregisterRequest": {
        "type": "object",
        "description": "Request body shared by all unregister routes.",
        "properties": {
          "schedule_time_ms": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Time in milliseconds when this request should be applied. Value `0` represents\ntime of the start request."
          }
        }
      },
...
    }
  }
}